backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
sqlparse==0.5.5
tzdata==2025.3
